use the uint32_t based crc API
This is an incompatible change! Databases created with an older version
can't be read anymore.

+++ src/database.h
@@ -11,7 +11,7 @@
 struct DBItem {
   char *kbuf;
   int kbuflen;
-  unsigned long crc;
+  uint32_t crc;
   struct DBItem *next;
 };
